Refactor parser interface
